Add files via upload
A Dynamic PIP page for Bitfocus Companion to control dynamic inputs in vMix 25.
Here is what it does - https://youtu.be/bb1UjEFIok0
The page is created in v. 2.2.0 (2.2.0+4160-beta-e04acafd) of Companion (recent beta of per days date.)

Credits goes to Matthew from Crazy-Logic on YouTube, since he was the main inspiration for this amazing surface!